+++ Assets/Project/Scripts/GameSession.cs
@@ -18,7 +18,7 @@ public class GameSession : MonoBehaviour {
     public const int DINNER_RUSH_START = 1730;
     public const int DINNER_RUSH_END = 2000;
 
-    public const int MAX_TIME_BTWN_ORDERS = 60;
+    public const int MAX_TIME_BTWN_ORDERS = 120;
     public const int MIN_TIME_BTWN_ORDERS = 10;
 
     public const float C_VAL = 3f;
@@ -28,6 +28,7 @@ public class GameSession : MonoBehaviour {
     private const int STATUS_LUNCH_RUSH = 2;
     private const int STATUS_DINNER_RUSH = 3;
     private const int STATUS_END_DAY = 4;
+    private const int STATUS_OVERTIME = 5;
 
     private const float LUNCH_DIFFICULTY_MOD = 1.32f;
     private const float DINNER_DIFFICULTY_MOD = 1.475f;
@@ -43,6 +44,9 @@ public class GameSession : MonoBehaviour {
     private int dayTime = 0; // hhmm format
 
     private int dayStatus = STATUS_NORMAL;
+
+    private float bVal;
+
 
     private int lastOrder = 0;
     private List<int> timeDiff;
@@ -67,6 +71,8 @@ public class GameSession : MonoBehaviour {
 
         numPlayers = info.numPlayers;
         spawnPlayers(info);
+
+        bVal = Mathf.Pow(MAX_TIME_BTWN_ORDERS * Mathf.Pow(0.75f, numPlayers - 1), C_VAL);
         StartCoroutine("startDay");
     }
 
@@ -78,7 +84,7 @@ public class GameSession : MonoBehaviour {
                 incrementTime((int)(minuteCount / MINUTE_VALUE));
                 minuteCount %= MINUTE_VALUE;
                 updateStatus();
-                if(dayStatus != STATUS_NO_ORDERS) {
+                if(dayStatus != STATUS_NO_ORDERS && dayStatus != STATUS_OVERTIME) {
                     randomChanceOrder();
                 }
             }
@@ -87,6 +93,7 @@ public class GameSession : MonoBehaviour {
     }
 
     private IEnumerator endDay() {
+        ticketBoard.resetTickets();
         daysCleared++;
         float sum = 0;
         foreach(int t in timeDiff) {
@@ -103,9 +110,11 @@ public class GameSession : MonoBehaviour {
         StartCoroutine("startDay");
     }
 
-    public IEnumerator finishOrder() {
-        score += Mathf.FloorToInt(100 + 1000 * difficulty * difficultyMod * (daysCleared + 1));
-        orderCompletionText.text = "Order complete!";
+    public IEnumerator finishOrder(int startTime) {
+        float timeRemaining = 1 - Mathf.Pow(1 - (float)(TicketGen.TICKET_DURATION - timeSpent(startTime, dayTime)) / TicketGen.TICKET_DURATION, 1.5f);
+        int pointVal = Mathf.FloorToInt(100 + 100 * difficulty * difficultyMod * timeRemaining);
+        score += pointVal;
+        orderCompletionText.text = "Order complete!\n+" + pointVal;
         yield return new WaitForSeconds(3f);
         orderCompletionText.text = "";
     }
@@ -113,7 +122,7 @@ public class GameSession : MonoBehaviour {
     private void spawnOrder() {
         timeDiff.Add(dayTime - lastOrder);
         lastOrder = dayTime;
-        ticketBoard.diffModifier = difficulty;
+        ticketBoard.diffModifier = difficulty * difficultyMod;
         ticketBoard.newOrder();
     }
 
@@ -182,7 +191,7 @@ public class GameSession : MonoBehaviour {
             float maxf = 0.5f;
             float floor = maxf * difficulty * difficultyMod;
             float rand = Random.Range(floor, 1f);
-            float catchup = (Mathf.Pow(last, C_VAL) / Mathf.Pow(MAX_TIME_BTWN_ORDERS, C_VAL)) * (1 - rand);
+            float catchup = (Mathf.Pow(last, C_VAL) / bVal) * (1 - rand);
             if((rand + catchup) > 0.984f) {
                 spawnOrder();
             }
@@ -199,24 +208,43 @@ public class GameSession : MonoBehaviour {
                     StartCoroutine(setStatus(STATUS_DINNER_RUSH));
                 } else if (dayTime >= END_ORDERS) {
                     StartCoroutine(setStatus(STATUS_NO_ORDERS));
+                } else if (dayTime == LUNCH_RUSH_START - 30) {
+                    StartCoroutine(displayMessage("Prepare for the Lunch Rush!", 2f));
+                } else if (dayTime == DINNER_RUSH_START - 30) {
+                    StartCoroutine(displayMessage("Prepare for the Dinner Rush!", 2f));
                 }
                 break;
             case STATUS_NO_ORDERS:
                 if (dayTime >= START_ORDERS && dayTime < END_ORDERS) {
+                    StartCoroutine(displayMessage("The Kitchen is now open!", 2f));
                     StartCoroutine(setStatus(STATUS_NORMAL));
                 } else if (dayTime >= END_DAY) {
-                    StartCoroutine(setStatus(STATUS_END_DAY));
-                    StartCoroutine("endDay");
+                    if (ticketBoard.numOrders() == 0) {
+                        StartCoroutine(setStatus(STATUS_END_DAY));
+                        StartCoroutine("endDay");
+                    } else {
+                        StartCoroutine(setStatus(STATUS_OVERTIME));
+                    }
                 }
                 break;
             case STATUS_LUNCH_RUSH:
                 if (dayTime >= LUNCH_RUSH_END) {
+                    StartCoroutine(displayMessage("Lunch Rush Completed", 2f));
                     StartCoroutine(setStatus(STATUS_NORMAL));
                 }
                 break;
             case STATUS_DINNER_RUSH:
                 if (dayTime >= DINNER_RUSH_END) {
+                    StartCoroutine(displayMessage("Dinner Rush Completed", 2f));
                     StartCoroutine(setStatus(STATUS_NORMAL));
+                }
+                break;
+            case STATUS_OVERTIME:
+                if(ticketBoard.numOrders() == 0) {
+                    StartCoroutine(setStatus(STATUS_END_DAY));
+                    StartCoroutine("endDay");
+                } else {
+                    score -= Mathf.FloorToInt(difficulty * 10);
                 }
                 break;
             default:
@@ -238,7 +266,11 @@ public class GameSession : MonoBehaviour {
         } else if (status == STATUS_DINNER_RUSH) {
             difficultyMod = DINNER_DIFFICULTY_MOD;
             yield return displayMessage("Dinner Rush!", 3f);
+        } else if (status == STATUS_OVERTIME) {
+            clockText.color = new Color(1f, 0f, 0f);
+            yield return displayMessage("Overtime!!", 3f);
         } else {
+            clockText.color = new Color(1f, 1f, 1f);
             difficultyMod = 1f;
         }
     }
@@ -257,5 +289,18 @@ public class GameSession : MonoBehaviour {
             lastOrder += 40;
         }
         return dayTime;
+    }
+
+    public int getTime()
+    {
+        return dayTime;
+    }
+
+    public static int timeSpent(int sTime, int curTime)
+    {
+        int sMin = sTime / 100 * 60 + sTime % 100;
+        int cMin = curTime / 100 * 60 + sTime % 100;
+
+        return cMin - sMin;
     }
 }